WIP towards PUBDEV-602: Fix AUC accuracy.

+++ h2o-core/src/main/java/hex/ConfusionMatrix.java
@@ -287,7 +287,7 @@ public class ConfusionMatrix extends Iced {
     String[] colType = new String[colHeader.length];
     String[] colFormat = new String[colHeader.length];
     for (int i=0; i<colFormat.length-1; ++i) {
-      colType[i]   = "integer";
+      colType[i]   = "long";
       colFormat[i] = "%d";
     }
     colType[colFormat.length-2]   = "double";
